# Music-platform/src/docs/api.yaml
openapi: 3.0.0
info:
  title: Music Platform backend service
  description: Specification for Music Platform backend service
  version: 1.0.0
servers:
  - url: http://localhost:8181/v1
paths:
...
  /playlist/:
    post:
      description: Get a certain playlist
      requestBody:
        content:
          application/json:
            schema:
              $ref: './components.yaml#/components/schemas/Playlist'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: inserted playlist's UUID
        400:
          $ref: './common/responses.yaml#/components/responses/400'
  /playlist/{playlistId}:
    get:
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Gets playlist
          content:
            application/json:
              schema:
                $ref: './components.yaml#/components/schemas/Playlist'
        400:
          $ref: './common/responses.yaml#/components/responses/400'
    delete:
      description: Delete user's playlist
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: string

      responses:
        200:
          description: OK
        400:
          $ref: './common/responses.yaml#/components/responses/400'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT